Execute each test file in a separate process (#525)
